Fixed various eclipse warnings

+++ src/ibis/zorilla/cluster/VivaldiService.java
@@ -77,7 +77,7 @@ public class VivaldiService implements Service, Runnable {
             }
 
             long time = end - start;
-            double rtt = (double) time / 1000000.0;
+            double rtt = time / 1000000.0;
 
             if (rtt < result) {
                 result = rtt;
